Rename abp.abp.swagger to volo.abp.swashbuckle
