+++ framework/src/Volo.Abp.Swashbuckle/Volo/Abp/Swashbuckle/AbpSwashbuckleModule.cs
@@ -1,16 +1,16 @@
 ﻿using Volo.Abp.Modularity;
 using Volo.Abp.VirtualFileSystem;
 
-namespace Volo.Abp
+namespace Volo.Abp.Swashbuckle
 {
     [DependsOn(typeof(AbpVirtualFileSystemModule))]
-    public class AbpSwaggerModule : AbpModule
+    public class AbpSwashbuckleModule : AbpModule
     {
         public override void ConfigureServices(ServiceConfigurationContext context)
         {
             Configure<AbpVirtualFileSystemOptions>(options =>
             {
-                options.FileSets.AddEmbedded<AbpSwaggerModule>();
+                options.FileSets.AddEmbedded<AbpSwashbuckleModule>();
             });
         }
     }
